All tests are passing with vertx-lib integration
Moving all the handleAuthorize code into its own class that implements RouterCreator has worked
well. All tests are passing, except one, which I'm not convinced was actually valid. It
wants to check for the absence of the algorithm, but relied on the some gymnastics to do so
(creating a test class inheriting from MainVerticle). I tried a few appraoches to this, none
of which have worked so far. Instead the exception isn't thrown as expected. I wonder if
the code on master only thinks that RuntimeException is being thrown where it is expecting it
to be thrown but instead is catching an irrelevant NPE instead that gets thrown later.

Going to switch into master to see if I can validate this.
