+++ src/main/java/org/folio/auth/authtokenmodule/MissingAlgorithmException.java
@@ -1,0 +1,7 @@
+package org.folio.auth.authtokenmodule;
+
+public class MissingAlgorithmException extends Exception {
+  public MissingAlgorithmException(String message, Exception cause) {
+    super(message, cause);
+  }
+}
